lib-level: improve API
The tile class equality is already checked by the property deserializer.
No need to check it twice. Tile size is already a contract between
the level design, art and level parsing. Let's make it available for
logic too. Same for the atlas margins and spacing as this is incredibly
useful information for the renderer. Finally, Transform was a poor name
choice for lib-level. It conflicts with Transform from lib-game.