lib-game: simplify text API
Get rid of the ECS world spawn API and replace it with simple
put_text. Also add API so the caller doesn't have to call format!
each time the string depends on some runtime value. This way
we will be able to optimize the rendering of the text to do
zero allocations.